fix(voice): review round — claim fencing, detached terminal writes, honest evaluation gates
Findings from the PR review, fixed at the invariant:
- An unclaimable-but-live build snoozes the job instead of succeeding —
  a crash inside the reclaim grace can no longer strand a row as running
  with no job left to work it; the sweep also re-offers stale queued rows.
- Every terminal transition (defer, fail, complete) runs on a detached
  context and is fenced to its claim generation via started_at, so a
  job timeout cannot lose the outcome and a reclaimed worker cannot win
  a terminal write over its replacement.
- The corpus snapshot hash now covers kind, register, and weight — a
  manifest edit reads as a changed corpus — with one spelling shared by
  create, claim, and complete; a corpus edited mid-run demotes the
  result to review instead of activating from an obsolete snapshot; the
  claim serializes hash and sample reads under the profile row lock.
- A deferred row is only claimable once its window reopens, and the
  deferral honors the router's exact budget boundary when it carries one.
- A malformed judge verdict or an unreadable predecessor blocks
  auto-activation instead of blending into a passing score; an unbound
  model lane fails as model_unavailable, not internal.
- Anti-AI floor: the German not-X-but-Y reframe is detected on its own
  wire, spaced numeric ranges keep their dashes, sentence stats respect
  ellipsis boundaries, and a document-sized first sample enters the
  prompt as a bounded excerpt.
