+++ backend/internal/compose/voicebuild.go
@@ -60,6 +60,14 @@ const voiceBuildDeferral = 6 * time.Hour
 // drafts and 5 judge calls, each small, with validator retry headroom.
 const voiceBuildTimeout = 10 * time.Minute
 
+// The live voice_build row states the worker distinguishes when a claim is
+// refused; a terminal row needs no further work.
+const (
+	voiceRowQueued   = "queued"
+	voiceRowDeferred = "deferred"
+	voiceRowRunning  = "running"
+)
+
 func voiceBuildInsertOpts() *river.InsertOpts {
 	return &river.InsertOpts{UniqueOpts: river.UniqueOpts{ByArgs: true, ByState: activeSweepStates}}
 }
@@ -146,28 +154,64 @@ func (w *voiceBuildWorker) Work(ctx context.Context, job *river.Job[VoiceBuildAr
 		return fmt.Errorf("voice_build %s: claim: %w", job.Args.BuildID, err)
 	}
 	if !claimed {
-		return nil
+		switch input.Build.Status {
+		case voiceRowQueued, voiceRowDeferred, voiceRowRunning:
+			// The row is still live but not claimable right now: a rival's
+			// fresh claim, a deferred window that has not opened, or a dead
+			// worker's row inside its reclaim grace. Succeeding here would
+			// strand the build with no job left to work it — come back.
+			return river.JobSnooze(2 * time.Minute)
+		default:
+			// Terminal — the row already owns its outcome.
+			return nil
+		}
 	}
+	// Terminal transitions run on a detached context: a job timeout must
+	// never cancel the write that records the outcome (deep-read precedent).
+	terminal, cancel := terminalCtx(ctx)
+	defer cancel()
+	claimedAt := claimTime(input)
 	if w.brain == nil {
-		return w.fail(ctx, buildID, "model_unavailable",
+		return w.fail(terminal, buildID, claimedAt, "model_unavailable",
 			"Voice building is unavailable until an AI provider is configured on the worker role.")
 	}
-	if err := w.run(ctx, buildID, input); err != nil {
+	if err := w.run(ctx, terminal, buildID, input); err != nil {
 		if errors.Is(err, ai.ErrBudgetDeferred) {
-			if deferErr := w.store.DeferBuild(ctx, buildID,
+			if deferErr := w.store.DeferBuild(terminal, buildID, claimedAt,
 				"The monthly AI budget is exhausted; the build resumes in the next window.",
-				w.now().Add(voiceBuildDeferral)); deferErr != nil {
+				w.deferralDeadline(err)); deferErr != nil {
 				return fmt.Errorf("voice_build %s: defer: %w", job.Args.BuildID, deferErr)
 			}
 			return nil
 		}
-		return w.fail(ctx, buildID, failureStatusCode(err), ai.SafeVoiceBuildFailure(err))
+		return w.fail(terminal, buildID, claimedAt, failureStatusCode(err), ai.SafeVoiceBuildFailure(err))
 	}
 	return nil
 }
 
-// run drives extract → evaluate → activate on a claimed build.
-func (w *voiceBuildWorker) run(ctx context.Context, buildID ids.UUID, input ai.VoiceBuildInput) error {
+// claimTime is the claim generation every terminal write fences on.
+func claimTime(input ai.VoiceBuildInput) time.Time {
+	if input.Build.StartedAt == nil {
+		// ClaimBuild always stamps started_at; a nil here would only mean a
+		// hand-edited row, and a zero fence simply never matches.
+		return time.Time{}
+	}
+	return *input.Build.StartedAt
+}
+
+// deferralDeadline honors the router's exact budget-window boundary when the
+// error carries one; the fixed fallback serves only a bare sentinel.
+func (w *voiceBuildWorker) deferralDeadline(err error) time.Time {
+	var deferral *ai.BudgetDeferralError
+	if errors.As(err, &deferral) && deferral.NextAttemptAt.After(w.now()) {
+		return deferral.NextAttemptAt
+	}
+	return w.now().Add(voiceBuildDeferral)
+}
+
+// run drives extract → evaluate → activate on a claimed build; terminal is
+// the detached context the completing write uses.
+func (w *voiceBuildWorker) run(ctx, terminal context.Context, buildID ids.UUID, input ai.VoiceBuildInput) error {
 	heldOut, buildSamples := splitVoiceHeldOut(input.Samples, input.Build.SourceHash)
 	if err := w.store.SetBuildStage(ctx, buildID, "extract"); err != nil {
 		w.log.WarnContext(ctx, "voice build stage update failed", "build", buildID.String(), "err", err)
@@ -192,7 +236,7 @@ func (w *voiceBuildWorker) run(ctx context.Context, buildID ids.UUID, input ai.V
 	if err := w.store.SetBuildStage(ctx, buildID, "activate"); err != nil {
 		w.log.WarnContext(ctx, "voice build stage update failed", "build", buildID.String(), "err", err)
 	}
-	_, err = w.store.CompleteBuild(ctx, buildID, ai.VoiceBuildOutcome{
+	_, err = w.store.CompleteBuild(terminal, buildID, claimTime(input), ai.VoiceBuildOutcome{
 		Artifact:       artifact,
 		Evaluation:     evaluated.Evaluation,
 		SampleDrafts:   evaluated.SampleDrafts,
@@ -209,19 +253,24 @@ func (w *voiceBuildWorker) run(ctx context.Context, buildID ids.UUID, input ai.V
 
 // fail records the terminal failure on the row; the job itself succeeds —
 // the row owns retry policy, not River.
-func (w *voiceBuildWorker) fail(ctx context.Context, buildID ids.UUID, statusCode, detail string) error {
-	if err := w.store.FailBuild(ctx, buildID, statusCode, detail); err != nil {
+func (w *voiceBuildWorker) fail(ctx context.Context, buildID ids.UUID, claimedAt time.Time, statusCode, detail string) error {
+	if err := w.store.FailBuild(ctx, buildID, claimedAt, statusCode, detail); err != nil {
 		return fmt.Errorf("voice_build %s: record failure: %w", buildID.String(), err)
 	}
 	w.log.WarnContext(ctx, "voice build failed", "build", buildID.String(), "status_code", statusCode, "detail", detail)
 	return nil
 }
 
-// failureStatusCode maps a build error class onto the row vocabulary:
-// malformed or unverifiable model output is invalid_output, everything else
-// is internal.
+// failureStatusCode maps a build error class onto the row vocabulary: a
+// missing or unbound model lane is model_unavailable, malformed or
+// unverifiable model output is invalid_output, everything else is internal.
 func failureStatusCode(err error) string {
 	text := err.Error()
+	for _, marker := range []string{"no model path", "no bound", "not bound", "unbound"} {
+		if strings.Contains(text, marker) {
+			return "model_unavailable"
+		}
+	}
 	for _, marker := range []string{"invalid JSON", "cited unknown sample", "is not verbatim", "output", "is empty"} {
 		if strings.Contains(text, marker) {
 			return "invalid_output"
