+++ backend/internal/modules/ai/antiai.go
@@ -19,7 +19,7 @@ type VoiceViolation struct {
 }
 
 var (
-	abstractContrast = regexp.MustCompile(`(?i)\b(?:it(?:'s| is)|this is|es (?:ist|geht))\s+not\b[^.!?]{0,100}\b(?:but|sondern)\b|\bnot about\b[^.!?]{0,100}\b(?:but|sondern)\b`)
+	abstractContrast = regexp.MustCompile(`(?i)\b(?:it(?:'s| is)|this is)\s+not\b[^.!?]{0,100}\bbut\b|\bnot about\b[^.!?]{0,100}\bbut\b|\b(?:es (?:ist|geht)|das ist)\s+nicht\b[^.!?]{0,100}\bsondern\b|\bnicht um\b[^.!?]{0,100}\bsondern\b`)
 	cannedOpener     = regexp.MustCompile(`(?i)^\s*(?:here(?:'s| is) the thing|the truth is|let(?:'s| us) be honest|die wahrheit ist|mal ehrlich)\b`)
 	genericCTA       = regexp.MustCompile(`(?i)\b(?:what do you think|agree\?|are you ready\?|is your (?:team|organization|unternehmen) ready|wie siehst du das\?|was denkst du\?)`)
 	aiEse            = regexp.MustCompile(`(?i)\b(?:delve|unlock|leverage|game[- ]changer|transformative|ever[- ]evolving|navigate the complexities|synergy|paradigm shift|ganzheitlich|bahnbrechend|in einer sich ständig wandelnden welt)\b`)
@@ -52,14 +52,27 @@ func containsParentheticalDash(text string) bool {
 		if r != '—' && r != '–' {
 			continue
 		}
-		previousDigit := i > 0 && runes[i-1] >= '0' && runes[i-1] <= '9'
-		nextDigit := i+1 < len(runes) && runes[i+1] >= '0' && runes[i+1] <= '9'
-		if previousDigit && nextDigit {
+		if isNumericRangeDash(runes, i) {
 			continue
 		}
 		return true
 	}
 	return false
+}
+
+// isNumericRangeDash exempts a dash whose nearest non-space neighbours are
+// digits: "2024–2026" and the spaced "2024 – 2026" are ranges, not asides.
+func isNumericRangeDash(runes []rune, i int) bool {
+	prev := i - 1
+	for prev >= 0 && runes[prev] == ' ' {
+		prev--
+	}
+	next := i + 1
+	for next < len(runes) && runes[next] == ' ' {
+		next++
+	}
+	return prev >= 0 && runes[prev] >= '0' && runes[prev] <= '9' &&
+		next < len(runes) && runes[next] >= '0' && runes[next] <= '9'
 }
 
 // SanitizeAIPatterns guarantees the hard punctuation rule after the rewrite
@@ -70,9 +83,7 @@ func SanitizeAIPatterns(text string) string {
 	for i := 0; i < len(runes); i++ {
 		r := runes[i]
 		if r == '—' || r == '–' {
-			previousDigit := i > 0 && runes[i-1] >= '0' && runes[i-1] <= '9'
-			nextDigit := i+1 < len(runes) && runes[i+1] >= '0' && runes[i+1] <= '9'
-			if previousDigit && nextDigit {
+			if isNumericRangeDash(runes, i) {
 				out = append(out, r)
 				continue
 			}
